soc: adi: max32: Add support for secondary risc-v core
Adds support for the secondary risc-v core on all applicable max32
family socs. The zero-riscy core supports the following instructions:

  - RV32I Base Integer Instruction Set
  - RV32E Base Integer Instruction Set (optional)
  - RV32C Standard Extension for Compressed Instructions
  - RV32M Integer Multiplication and Division Instruction Set Extension
    (optional)

We need to determine if the optional RV32E and RV32M were enabled or
disabled when the zero-riscy core was instantiated on the max32 family.

The zero-riscy core does not implement the full risc-v privileged
specification, so we need to revisit whether it's valid to select
CONFIG_RISCV_PRIVILEGED when we enable proper support for interrupts.

See https://pulp-platform.org/docs/user_manual.pdf

Signed-off-by: Maureen Helm <maureen.helm@analog.com>